+++ packages/qwik/src/validation/maxRange.ts
@@ -1,17 +1,17 @@
 import { $, type QRL } from '@builder.io/qwik';
 
-type Value = string | null | undefined;
+type Value = string | Date | null | undefined;
 
 /**
- * Creates a validation functions that validates the range of a string.
+ * Creates a validation functions that validates the range of a string or date.
  *
- * @param requirement The maximum string range.
+ * @param requirement The maximum range.
  * @param error The error message.
  *
  * @returns A validation function.
  */
 export function maxRange(
-  requirement: string,
+  requirement: string | Date,
   error: string
 ): QRL<(value: Value) => string> {
   return $((value: Value) => (value && value > requirement ? error : ''));