+++ playgrounds/qwik/src/components/TextInput.tsx
@@ -1,5 +1,7 @@
 import {
   component$,
+  useSignal,
+  useTask$,
   type PropFunction,
   type QwikChangeEvent,
   type QwikFocusEvent,
@@ -36,8 +38,14 @@ type TextInputProps = {
  * displayed in or around the field to communicate the entry requirements.
  */
 export const TextInput = component$(
-  ({ label, error, ...props }: TextInputProps) => {
+  ({ label, value, error, ...props }: TextInputProps) => {
     const { name, required } = props;
+    const input = useSignal<string | number>();
+    useTask$(({ track }) => {
+      if (!Number.isNaN(track(() => value))) {
+        input.value = value;
+      }
+    });
     return (
       <div class={clsx('px-8 lg:px-10', props.class)}>
         <InputLabel name={name} label={label} required={required} />
@@ -50,6 +58,7 @@ export const TextInput = component$(
               : 'border-slate-200 hover:border-slate-300 focus:border-sky-600/50 dark:border-slate-800 dark:hover:border-slate-700 dark:focus:border-sky-400/50'
           )}
           id={name}
+          value={input.value}
           aria-invalid={!!error}
           aria-errormessage={`${name}-error`}
         />