+++ playgrounds/qwik/src/routes/actions/special/index.tsx
@@ -117,7 +117,7 @@ export default component$(() => {
           console.log(values);
         }}
       >
-        <Field name="number">
+        <Field name="number" type="number">
           {(field, props) => (
             <TextInput
               {...props}
@@ -129,7 +129,7 @@ export default component$(() => {
           )}
         </Field>
 
-        <Field name="range">
+        <Field name="range" type="number">
           {(field, props) => (
             <Slider
               {...props}
@@ -150,7 +150,7 @@ export default component$(() => {
             { label: 'Option 2', value: 'option_2' },
             { label: 'Option 3', value: 'option_3' },
           ].map(({ label, value }) => (
-            <Field key={value} name="checkbox.array">
+            <Field key={value} name="checkbox.array" type="string[]">
               {(field, props) => (
                 <Checkbox
                   {...props}
@@ -165,7 +165,7 @@ export default component$(() => {
           ))}
         </div>
 
-        <Field name="checkbox.boolean">
+        <Field name="checkbox.boolean" type="boolean">
           {(field, props) => (
             <Checkbox
               {...props}
@@ -176,7 +176,7 @@ export default component$(() => {
           )}
         </Field>
 
-        <Field name="select.array">
+        <Field name="select.array" type="string[]">
           {(field, props) => (
             <Select
               {...props}
@@ -209,7 +209,7 @@ export default component$(() => {
           )}
         </Field>
 
-        <Field name="file.list">
+        <Field name="file.list" type="File[]">
           {(field, props) => (
             <FileInput
               {...props}
@@ -221,7 +221,7 @@ export default component$(() => {
           )}
         </Field>
 
-        <Field name="file.item">
+        <Field name="file.item" type="File">
           {(field, props) => (
             <FileInput
               {...props}